Add VALIDATED state evidence (run-000004)
- Acceptance gates evaluated against schema v2.0.0
- G1-G4, G6: PASS (5/6 gates)
- G5 (Determinism): PENDING - requires run-000005
- Overall: CONDITIONAL_PASS
- Validator: RawrXDValidator.exe v1.0.0

Gate Results:
- G1 Source Identity: PASS (commit 857610125)
- G2 Binary Integrity: PASS (SHA256 verified)
- G3 Execution Completion: PASS (exit code 0)
- G4 Numerical Correctness: PASS (max_error=0.0)
- G5 Determinism: PENDING (needs repeat run)
- G6 Evidence Closure: PASS (all artifacts present)

Lifecycle transition: EXECUTED -> VALIDATED
Next: VERIFIED state with determinism proof

+++ src/integration/val016_source_repair_v2.cpp
@@ -122,7 +122,7 @@ struct RepairPatch {
         j["offset"] = offset;
         j["before"] = before;
         j["after"] = after;
-        j["confidence"] = static_cast<double>(confidence);
+        j["confidence"] = confidence;
         j["reason"] = reason;
         return j;
     }
